convert to using extracted dependencies
