--- populus/contracts/__init__.py
@@ -1,16 +1,0 @@
-# NOQA
-from populus.contracts.utils import (  # NOQA
-    deploy_contract,
-    get_linker_dependencies,
-    link_contract_dependency,
-    strip_0x_prefix,
-)
-from populus.contracts.functions import (  # NOQA
-    Function,
-    FunctionGroup,
-)
-from populus.contracts.events import Event  # NOQA
-from populus.contracts.core import (  # NOQA
-    Contract,
-    package_contracts,
-)
