+++ requirements.txt
@@ -1,5 +1,6 @@
 click>=5.1
 ethereum-rpc-client>=0.4.0
+ethereum-contract>=0.1.0
 ethereum>=1.0.3
 eth-testrpc>=0.1.17
 pytest>=2.7.2
