+++ tests/deployment/test_finding_linker_dependencies.py
@@ -1,16 +1,14 @@
 import os
 
 from populus.utils import load_contracts
-from populus.contracts import (
+from populus.deployment import (
     package_contracts,
-)
-from populus.contracts.utils import (
     get_linker_dependencies,
 )
 
 
 PROJECTS_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'projects')
-project_dir = os.path.join(PROJECTS_DIR, 'test-03')
+project_dir = os.path.join(PROJECTS_DIR, 'test-01')
 
 
 def test_extracting_linker_dependencies():
